Motley/requirements.txt
aws-cdk-lib>=2.140.0
constructs>=10.0.0,<11.0.0
boto3
aws-cdk.lambda-layer-kubectl-v23
aws-cdk.lambda-layer-kubectl-v24
aws-cdk.aws-sagemaker-alpha
aws-cdk.aws-synthetics-alpha
aws-cdk.aws_lambda_python_alpha